app/requirements.txt
Django==1.8.4
PIL==1.1.7
Pillow==2.9.0
argparse==1.2.1
django-appconf==1.0.1
django-simple-captcha==0.4.6
lockfile==0.10.2
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
sorl-thumbnail==12.3
wsgiref==0.1.2
